requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
# created-by: conda 24.11.3
absl-py=2.2.2=pypi_0
astunparse=1.6.3=pypi_0
blas=1.0=mkl
brotli-python=1.0.9=py312h5da7b33_9
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2025.2.25=haa95532_0
cachetools=5.5.2=pypi_0
certifi=2025.4.26=py312haa95532_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
colorama=0.4.6=pypi_0
contourpy=1.3.2=pypi_0
cuda-cccl=12.9.27=0
cuda-cccl_win-64=12.9.27=0
cuda-cudart=11.8.89=0
cuda-cudart-dev=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-libraries-dev=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvrtc-dev=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-profiler-api=12.9.19=0
cuda-runtime=11.8.0=0
cuda-version=12.9=3
cycler=0.12.1=pypi_0
decorator=5.2.1=pypi_0
dill=0.4.0=pypi_0
expat=2.7.1=h8ddb27b_0
filelock=3.17.0=py312haa95532_0
flatbuffers=25.2.10=pypi_0
fonttools=4.57.0=pypi_0
freeglut=3.4.0=hd77b12b_0
freetype=2.13.3=h0620614_0
fsspec=2025.3.2=pypi_0
gast=0.6.0=pypi_0
giflib=5.2.2=h7edc060_0
gmp=6.3.0=h537511b_0
gmpy2=2.2.1=py312h827c3e9_0
google-auth=2.39.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.71.0=pypi_0
h5py=3.13.0=pypi_0
idna=3.7=py312haa95532_0
intel-openmp=2023.1.0=h59b6b97_46320
jinja2=3.1.6=py312haa95532_0
joblib=1.4.2=pypi_0
jpeg=9e=h827c3e9_3
keras=3.9.2=pypi_0
kiwisolver=1.4.8=pypi_0
lazy-loader=0.4=pypi_0
lcms2=2.16=h62be587_1
lerc=4.0.0=h5da7b33_0
libclang=18.1.1=pypi_0
libcublas=11.11.3.6=0
libcublas-dev=11.11.3.6=0
libcufft=10.9.0.58=0
libcufft-dev=10.9.0.58=0
libcurand=10.3.10.19=0
libcurand-dev=10.3.10.19=0
libcusolver=11.4.1.48=0
libcusolver-dev=11.4.1.48=0
libcusparse=11.7.5.86=0
libcusparse-dev=11.7.5.86=0
libdeflate=1.22=h5bf469e_0
libffi=3.4.4=hd77b12b_1
libjpeg-turbo=2.0.0=h196d8e1_0
libnpp=11.8.0.86=0
libnpp-dev=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libnvjpeg-dev=11.9.0.86=0
libpng=1.6.39=h8cc25b3_0
libtiff=4.7.0=h404307b_0
libuv=1.48.0=h827c3e9_0
libwebp=1.3.2=h18467be_1
libwebp-base=1.3.2=h3d04722_1
lz4-c=1.9.4=h2bbff1b_1
markdown=3.8=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=3.0.2=py312h827c3e9_0
matplotlib=3.10.1=pypi_0
mdurl=0.1.2=pypi_0
mkl=2023.1.0=h6b88ed4_46358
mkl-service=2.4.0=py312h827c3e9_2
mkl_fft=1.3.11=py312h827c3e9_0
mkl_random=1.2.8=py312h0158946_0
ml-dtypes=0.5.1=pypi_0
mne=1.9.0=pypi_0
mpc=1.3.1=h827c3e9_0
mpfr=4.2.1=h56c3642_0
mpmath=1.3.0=py312haa95532_0
namex=0.0.9=pypi_0
networkx=3.4.2=py312haa95532_0
nibabel=5.3.2=pypi_0
numpy=1.26.4=pypi_0
oauthlib=3.2.2=pypi_0
openjpeg=2.5.2=h9b5d1b5_1
openssl=3.0.16=h3f729d1_0
opt-einsum=3.4.0=pypi_0
optree=0.15.0=pypi_0
packaging=25.0=pypi_0
pandarallel=1.6.5=pypi_0
pandas=1.5.3=pypi_0
pillow=11.1.0=py312hea0d53e_1
pip=25.1=pyhc872135_1
platformdirs=4.3.7=pypi_0
pooch=1.8.2=pypi_0
protobuf=3.20.3=pypi_0
psutil=7.0.0=pypi_0
pyasn1=0.6.1=pypi_0
pyasn1-modules=0.4.2=pypi_0
pygments=2.19.1=pypi_0
pyhealth=1.1.6=pypi_0
pyparsing=3.2.3=pypi_0
pysocks=1.7.1=py312haa95532_0
python=3.12.9=h14ffc60_0
python-dateutil=2.9.0.post0=pypi_0
pytorch=2.5.1=py3.12_cuda11.8_cudnn9_0
pytorch-cuda=11.8=h24eeafa_6
pytorch-mutex=1.0=cuda
pytz=2025.2=pypi_0
pyyaml=6.0.2=py312h827c3e9_0
rdkit=2024.9.6=pypi_0
requests=2.32.3=py312haa95532_1
requests-oauthlib=2.0.0=pypi_0
rich=14.0.0=pypi_0
rsa=4.9.1=pypi_0
scikit-learn=1.3.2=pypi_0
scipy=1.15.2=pypi_0
setuptools=72.1.0=py312haa95532_0
six=1.17.0=pypi_0
sqlite=3.45.3=h2bbff1b_0
sympy=1.13.1=pypi_0
tbb=2021.8.0=h59b6b97_0
tensorboard=2.19.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
termcolor=3.1.0=pypi_0
threadpoolctl=3.6.0=pypi_0
tk=8.6.14=h0416ee5_0
torchaudio=2.5.1=pypi_0
torchvision=0.20.1=pypi_0
tqdm=4.67.1=pypi_0
typing_extensions=4.12.2=py312haa95532_0
tzdata=2025.2=pypi_0
urllib3=1.26.15=pypi_0
vc=14.42=haa95532_5
vs2015_runtime=14.42.34433=hbfb602d_5
werkzeug=3.1.3=pypi_0
wheel=0.45.1=py312haa95532_0
win_inet_pton=1.1.0=py312haa95532_0
wrapt=1.17.2=pypi_0
xz=5.6.4=h4754444_1
yaml=0.2.5=he774522_0
zlib=1.2.13=h8cc25b3_1
zstd=1.5.6=h8880b57_0
